overhaul macros; windows, move Chuck_IO to chuck_io.h/cpp
