+++ src/core/chuck_oo.cpp
@@ -40,53 +40,12 @@
 #include "util_math.h"
 
 #include <algorithm>
-#include <cstdint> // 1.5.0.1 (ge) added; requires c++11
-#include <iomanip>
 #include <iostream>
-#include <sstream>
-#include <typeinfo>
 using namespace std;
-
-#if defined(__PLATFORM_WINDOWS__)
-  #include "dirent_win32.h"
-#endif
 
 
 // initialize
 t_CKBOOL Chuck_VM_Object::our_locks_in_effect = TRUE;
-
-// constants
-const t_CKINT Chuck_IO::TYPE_ASCII  = 0x1;
-const t_CKINT Chuck_IO::TYPE_BINARY = 0x2;
-const t_CKINT Chuck_IO::FLOAT32 = 0x10;
-const t_CKINT Chuck_IO::FLOAT64 = 0x20;
-const t_CKINT Chuck_IO::INT8  = 0x100;
-const t_CKINT Chuck_IO::INT16 = 0x200;
-const t_CKINT Chuck_IO::INT24 = 0x400;
-const t_CKINT Chuck_IO::INT32 = 0x800;
-const t_CKINT Chuck_IO::INT64 = 0x1000;
-const t_CKINT Chuck_IO::SINT8 = 0x2000;
-const t_CKINT Chuck_IO::SINT16 = 0x4000;
-const t_CKINT Chuck_IO::SINT24 = 0x8000;
-const t_CKINT Chuck_IO::SINT32 = 0x10000;
-const t_CKINT Chuck_IO::SINT64 = 0x20000;
-const t_CKINT Chuck_IO::UINT8  = 0x40000;
-const t_CKINT Chuck_IO::UINT16 = 0x80000;
-const t_CKINT Chuck_IO::UINT24 = 0x100000;
-const t_CKINT Chuck_IO::UINT32 = 0x200000;
-const t_CKINT Chuck_IO::UINT64 = 0x400000;
-const t_CKINT Chuck_IO::FLAG_READONLY = 0x100;
-const t_CKINT Chuck_IO::FLAG_WRITEONLY = 0x200;
-const t_CKINT Chuck_IO::FLAG_READ_WRITE = 0x400;
-const t_CKINT Chuck_IO::FLAG_APPEND = 0x800;
-
-#ifndef __DISABLE_THREADS__
-const t_CKINT Chuck_IO::MODE_SYNC = 0;
-const t_CKINT Chuck_IO::MODE_ASYNC = 1;
-#else
-const t_CKINT Chuck_IO::MODE_SYNC = 1;
-const t_CKINT Chuck_IO::MODE_ASYNC = 0;
-#endif
 
 
 
@@ -2312,7 +2271,7 @@ void Chuck_Array32::shuffle()
 
 
 
-// static
+// Chuck_Event static instantiation
 t_CKUINT Chuck_Event::our_can_wait = 0;
 
 //-----------------------------------------------------------------------------
@@ -2844,1392 +2803,4 @@ void Chuck_Event::wait( Chuck_VM_Shred * shred, Chuck_VM * vm )
         t_CKTIME *& sp = (t_CKTIME *&)shred->reg->sp;
         push_( sp, shred->now );
     }
-}
-
-
-
-
-//-----------------------------------------------------------------------------
-// name: Chuck_IO Constructor
-// desc: Empty because you cannot construct a Chuck_IO object
-//-----------------------------------------------------------------------------
-Chuck_IO::Chuck_IO() : m_asyncEvent(NULL)
-{
-#ifndef __DISABLE_THREADS__
-    m_thread = NULL;
-#endif
-}
-
-
-
-
-//-----------------------------------------------------------------------------
-// name: Chuck_IO Destructor
-// desc: Empty becuase you cannot destruct a Chuck_IO object
-//-----------------------------------------------------------------------------
-Chuck_IO::~Chuck_IO()
-{ }
-
-
-
-
-// #ifndef __DISABLE_FILEIO__
-//-----------------------------------------------------------------------------
-// name: Chuck_IO_File()
-// desc: constructor
-//-----------------------------------------------------------------------------
-Chuck_IO_File::Chuck_IO_File( Chuck_VM * vm )
-{
-    m_vmRef = vm;
-    // zero things out
-    m_flags = 0;
-    m_iomode = MODE_SYNC;
-    m_path = "";
-    m_dir = NULL;
-    m_dir_start = 0;
-    m_asyncEvent = new Chuck_Event;
-    initialize_object( m_asyncEvent, vm->env()->t_event );
-    #ifndef __DISABLE_THREADS__
-    m_thread = new XThread;
-    #endif
-}
-
-
-
-
-//-----------------------------------------------------------------------------
-// name: ~Chuck_IO_File()
-// desc: destructor
-//-----------------------------------------------------------------------------
-Chuck_IO_File::~Chuck_IO_File()
-{
-    // clean up
-    this->close();
-    delete m_asyncEvent;
-    #ifndef __DISABLE_THREADS__
-    delete m_thread;
-    #endif
-}
-
-
-
-
-//-----------------------------------------------------------------------------
-// name: open
-// desc: open file from disk
-//-----------------------------------------------------------------------------
-t_CKBOOL Chuck_IO_File::open( const string & path, t_CKINT flags )
-{
-    // log
-    EM_log( CK_LOG_INFO, "FileIO: opening file from disk..." );
-    EM_log( CK_LOG_INFO, "FileIO: path: %s", path.c_str() );
-    EM_pushlog();
-
-    // if no flag specified, make it READ by default
-    if( !(flags & FLAG_READ_WRITE) &&
-        !(flags & FLAG_READONLY) &&
-        !(flags & FLAG_WRITEONLY) &&
-        !(flags & FLAG_APPEND) )
-    {
-        flags |= FLAG_READONLY;
-    }
-
-    // if both read and write, enable read and write
-    if( flags & FLAG_READONLY && flags & FLAG_WRITEONLY )
-    {
-        flags ^= FLAG_READONLY;
-        flags ^= FLAG_WRITEONLY;
-        flags |= FLAG_READ_WRITE;
-    }
-
-    // set open flags
-    ios_base::openmode theMode;
-
-    // check flags for errors
-    if ((flags & TYPE_ASCII) &&
-        (flags & TYPE_BINARY))
-    {
-        EM_error3( "[chuck](via FileIO): cannot open file in both ASCII and binary mode" );
-        goto error;
-    }
-
-    if ((flags & FLAG_READ_WRITE) &&
-        (flags & FLAG_READONLY))
-    {
-        EM_error3( "[chuck](via FileIO): conflicting flags: READ_WRITE and READ" );
-        goto error;
-    }
-
-    if ((flags & FLAG_READ_WRITE) &&
-        (flags & FLAG_WRITEONLY))
-    {
-        EM_error3( "[chuck](via FileIO): conflicting flags: READ_WRITE and WRITE" );
-        goto error;
-    }
-
-    if ((flags & FLAG_READ_WRITE) &&
-        (flags & FLAG_APPEND))
-    {
-        EM_error3( "[chuck](via FileIO): conflicting flags: READ_WRITE and APPEND" );
-        goto error;
-    }
-
-    if ((flags & FLAG_WRITEONLY) &&
-        (flags & FLAG_READONLY))
-    {
-        EM_error3( "[chuck](via FileIO): conflicting flags: WRITE and READ" );
-        goto error;
-    }
-
-    if ((flags & FLAG_APPEND) &&
-        (flags & FLAG_READONLY))
-    {
-        EM_error3( "[chuck](via FileIO): conflicting flags: APPEND and FLAG_READ" );
-        goto error;
-    }
-
-    if (flags & FLAG_READ_WRITE)
-        theMode = ios_base::in | ios_base::out;
-    else if (flags & FLAG_READONLY)
-        theMode = ios_base::in;
-    else if (flags & FLAG_APPEND)
-        theMode = ios_base::out | ios_base::app;
-    else if (flags & FLAG_WRITEONLY)
-        theMode = ios_base::out | ios_base::trunc;
-
-    if (flags & TYPE_BINARY)
-        theMode |= ios_base::binary;
-
-    // close first
-    if( m_io.is_open() )
-        this->close();
-
-    // try to open as a dir first (fixed 1.3.0.0 removed warning)
-    m_dir = opendir(path.c_str());
-    if( m_dir )
-    {
-        EM_poplog();
-        return TRUE;
-    }
-
-    // not a dir, create file if it does not exist unless flag is
-    // readonly
-    if( !(flags & FLAG_READONLY) )
-    {
-        m_io.open( path.c_str(), ios_base::in );
-        if ( m_io.fail() )
-        {
-            m_io.clear();
-            m_io.open( path.c_str(), ios_base::out | ios_base::trunc );
-            m_io.close();
-        }
-        else
-            m_io.close();
-    }
-
-    //open file
-    m_io.open( path.c_str(), theMode );
-
-    // seek to beginning if necessary
-    if (flags & FLAG_READ_WRITE)
-    {
-        m_io.seekp(0);
-        m_io.seekg(0);
-    }
-
-    /* ATODO: Ge's code
-     // windows sucks for being creative in the wrong places
-     #ifdef __PLATFORM_WINDOWS__
-     // if( flags ^ Chuck_IO::TRUNCATE && flags | Chuck_IO::READ ) nMode |= ios::nocreate;
-     m_io.open( path.c_str(), nMode );
-     #else
-     m_io.open( path.c_str(), (_Ios_Openmode)nMode );
-     #endif
-     */
-
-    // check for error
-    if( !(m_io.is_open()) )
-    {
-        // EM_error3( "[chuck](via FileIO): cannot open file: '%s'", path.c_str() );
-        goto error;
-    }
-
-    // set path
-    m_path = path;
-    // set flags
-    m_flags = flags;
-    if (!(flags & TYPE_BINARY))
-        m_flags |= Chuck_IO_File::TYPE_ASCII; // ASCII is default
-    // set mode
-    m_iomode = MODE_SYNC;
-
-    // pop
-    EM_poplog();
-
-    return TRUE;
-
-error:
-
-    // pop
-    EM_poplog();
-
-    // reset
-    m_path = "";
-    m_flags = 0;
-    m_iomode = MODE_SYNC;
-    m_io.clear();
-    m_io.close();
-
-    return FALSE;
-}
-
-
-
-
-//-----------------------------------------------------------------------------
-// name: close
-// desc: close file
-//-----------------------------------------------------------------------------
-void Chuck_IO_File::close()
-{
-    // log
-    EM_log( CK_LOG_INFO, "FileIO: closing file '%s'...", m_path.c_str() );
-    // close it
-    m_io.close();
-    m_flags = 0;
-    m_path = "";
-    m_iomode = Chuck_IO::MODE_SYNC;
-    if ( m_dir ) {
-        closedir( m_dir );
-        m_dir = NULL;
-        m_dir_start = 0;
-    }
-}
-
-
-
-
-//-----------------------------------------------------------------------------
-// name: good()
-// desc: ...
-//-----------------------------------------------------------------------------
-t_CKBOOL Chuck_IO_File::good()
-{
-    return m_dir || m_io.is_open();
-    // return (!m_dir) && (!m_io.is_open());
-}
-
-
-
-
-//-----------------------------------------------------------------------------
-// name: flush()
-// desc: ...
-//-----------------------------------------------------------------------------
-void Chuck_IO_File::flush()
-{
-    // sanity
-    if ( m_dir )
-    {
-        EM_error3( "[chuck](via FileIO): cannot flush on directory" );
-        return;
-    }
-    m_io.flush();
-}
-
-
-
-
-//-----------------------------------------------------------------------------
-// name: mode()
-// desc: ...
-//-----------------------------------------------------------------------------
-t_CKINT Chuck_IO_File::mode()
-{
-    // sanity
-    if ( m_dir )
-    {
-        EM_error3( "[chuck](via FileIO): cannot get mode on directory" );
-        return -1;
-    }
-    return m_iomode;
-}
-
-
-
-
-//-----------------------------------------------------------------------------
-// name: mode( t_CKINT flag )
-// desc: ...
-//-----------------------------------------------------------------------------
-void Chuck_IO_File::mode( t_CKINT flag )
-{
-    // sanity
-    if ( m_dir )
-    {
-        EM_error3( "[chuck](via FileIO): cannot set mode on directory" );
-        return;
-    }
-    if ( (flag != Chuck_IO::MODE_ASYNC) && (flag != Chuck_IO::MODE_SYNC) )
-    {
-        EM_error3( "[chuck](via FileIO): invalid mode flag" );
-        return;
-    }
-
-    m_iomode = flag;
-}
-
-
-
-
-//-----------------------------------------------------------------------------
-// name: size()
-// desc: Returns the size of the file in bytes, or -1 if no file is opened or
-//       if a directory is opened.
-//-----------------------------------------------------------------------------
-t_CKINT Chuck_IO_File::size()
-{
-    if (!(m_io.is_open())) return -1;
-    if ( m_dir )
-    {
-        EM_error3( "[chuck](via FileIO): cannot get size on a directory" );
-        return -1;
-    }
-
-    // no easy way to find file size in C++
-    // have to seek to end, report position
-    FILE * stream = fopen( m_path.c_str(), "r" );
-    fseek( stream, 0L, SEEK_END );
-    long endPos = ftell( stream );
-    fclose( stream );
-    return endPos;
-}
-
-
-
-
-//-----------------------------------------------------------------------------
-// name: seek( t_CKINT pos )
-// desc: Seeks to the specified byte offset in the file.
-//-----------------------------------------------------------------------------
-void Chuck_IO_File::seek( t_CKINT pos )
-{
-    if ( !(m_io.is_open()) )
-    {
-        EM_error3( "[chuck](via FileIO): cannot seek: no file is open" );
-        return;
-    }
-    if ( m_dir )
-    {
-        EM_error3( "[chuck](via FileIO): cannot seek on a directory" );
-        return;
-    }
-    // 1.5.0.0 (ge) | added since seekg fails if EOF already reached
-    m_io.clear();
-    m_io.seekg( pos );
-    m_io.seekp( pos );
-}
-
-
-
-
-//-----------------------------------------------------------------------------
-// name: tell()
-// desc: Returns the byte offset into the file, or -1 if no file is opened.
-//-----------------------------------------------------------------------------
-t_CKINT Chuck_IO_File::tell()
-{
-    if (!(m_io.is_open()))
-        return -1;
-    if ( m_dir )
-    {
-        EM_error3( "[chuck](via FileIO): cannot tell on directory" );
-        return -1;
-    }
-
-    return m_io.tellg();
-}
-
-
-
-
-//-----------------------------------------------------------------------------
-// name: isDir()
-// desc: ...
-//-----------------------------------------------------------------------------
-t_CKINT Chuck_IO_File::isDir()
-{
-    return m_dir != NULL;
-}
-
-
-
-
-//-----------------------------------------------------------------------------
-// name: dirList()
-// desc: ...
-//-----------------------------------------------------------------------------
-Chuck_Array4 * Chuck_IO_File::dirList()
-{
-    // sanity
-    if ( !m_dir )
-    {
-        EM_error3( "[chuck](via FileIO): cannot get list: no directory open" );
-        Chuck_Array4 *ret = new Chuck_Array4( TRUE, 0 );
-        initialize_object( ret, m_vmRef->env()->t_array );
-        return ret;
-    }
-
-    // fill vector with entry names
-    rewinddir( m_dir );
-    std::vector<Chuck_String *> entrylist;
-
-    // 1.5.0.0 (ge) | #chunreal UE-forced refactor
-    struct dirent * ent = readdir( m_dir );
-    // was: while( (ent = readdir( m_dir ) ) <-- in ge's opinion this is cleaner splitting readir() into two places
-    while( ent != NULL ) // fixed 1.3.0.0: removed warning
-    {
-        // pass NULL as shred ref
-        Chuck_String *s = (Chuck_String *)instantiate_and_initialize_object( m_vmRef->env()->t_string, NULL, m_vmRef );
-        s->set( std::string( ent->d_name ) );
-        if ( s->str() != ".." && s->str() != "." )
-        {
-            // don't include .. and . in the list
-            entrylist.push_back( s );
-        }
-        // #chunreal refactor
-        ent = readdir( m_dir );
-    }
-
-    // make array
-    Chuck_Array4 *array = new Chuck_Array4( true, entrylist.size() );
-    initialize_object( array, m_vmRef->env()->t_array );
-    for ( int i = 0; i < entrylist.size(); i++ )
-        array->set( i, (t_CKUINT) entrylist[i] );
-    return array;
-}
-
-
-
-
-//-----------------------------------------------------------------------------
-// name: read( t_CKINT length )
-// desc: ...
-//-----------------------------------------------------------------------------
-/*Chuck_String * Chuck_IO_File::read( t_CKINT length )
-{
-    // sanity
-    if (!(m_io.is_open())) {
-        EM_error3( "[chuck](via FileIO): cannot read: no file open" );
-        return new Chuck_String( "" );
-    }
-
-    if (m_io.fail()) {
-        EM_error3( "[chuck](via FileIO): cannot read: I/O stream failed" );
-        return new Chuck_String( "" );
-    }
-
-    if ( m_dir )
-    {
-        EM_error3( "[chuck](via FileIO): cannot read on a directory" );
-        return new Chuck_String( "" );
-    }
-
-    char buf[length+1];
-    m_io.read( buf, length );
-    buf[m_io.gcount()] = '\0';
-    string s( buf );
-    return new Chuck_String( s );
-}*/
-
-
-
-
-//-----------------------------------------------------------------------------
-// name: readLine()
-// desc: read line
-//-----------------------------------------------------------------------------
-Chuck_String * Chuck_IO_File::readLine()
-{
-    // sanity
-    if (!(m_io.is_open())) {
-        EM_error3( "[chuck](via FileIO): cannot readLine: no file open" );
-        return new Chuck_String( "" );
-    }
-
-    if (m_io.fail()) {
-        EM_error3( "[chuck](via FileIO): cannot readLine: I/O stream failed" );
-        return new Chuck_String( "" );
-    }
-
-    if ( m_dir )
-    {
-        EM_error3( "[chuck](via FileIO): cannot readLine on directory" );
-        return new Chuck_String( "" );
-    }
-
-    string s;
-    getline( m_io, s );
-
-    // chuck str
-    Chuck_String * str = new Chuck_String( s );
-    // initialize | 1.5.0.0 (ge) | added initialize_object
-    initialize_object( str, m_vmRef->env()->t_string );
-
-    // note this chuck string still needs to be initialized
-    return str;
-}
-
-
-
-
-//-----------------------------------------------------------------------------
-// name: readInt( t_CKINT flags )
-// desc: ...
-//-----------------------------------------------------------------------------
-t_CKINT Chuck_IO_File::readInt( t_CKINT flags )
-{
-    // sanity
-    if( !(m_io.is_open()) ) {
-        EM_error3( "[chuck](via FileIO): cannot readInt: no file open" );
-        return 0;
-    }
-
-    if( m_io.eof() ) {
-        EM_error3( "[chuck](via FileIO): cannot readInt: EOF reached" );
-        return 0;
-    }
-
-    if( m_dir )
-    {
-        EM_error3( "[chuck](via FileIO): cannot read on directory" );
-        return 0;
-    }
-
-    if( m_io.fail() ) {
-        EM_error3( "[chuck](via FileIO): cannot readInt: I/O stream failed" );
-        return 0;
-    }
-
-    // check mode ASCII vs. binary
-    if( m_flags & TYPE_ASCII )
-    {
-        // ASCII
-        t_CKINT val = 0;
-        m_io >> val;
-        // if (m_io.fail())
-        //     EM_error3( "[chuck](via FileIO): cannot readInt: I/O stream failed" );
-        return val;
-    }
-    else if( m_flags & TYPE_BINARY )
-    {
-        // binary
-        if( flags & Chuck_IO::INT8 || flags & Chuck_IO::UINT8 ) // 1.5.0.1 (ge) added UINT
-        {
-            // unsigned 8-bit
-            uint8_t i; // 1.5.0.1 (ge) changed to int8_t from unsigned char
-            m_io.read( (char *)&i, 1 );
-            if( m_io.gcount() != 1 )
-                EM_error3( "[chuck](via FileIO): cannot readInt: not enough bytes left" );
-            else if( m_io.fail() )
-                EM_error3( "[chuck](via FileIO): cannot readInt: I/O stream failed" );
-            return (t_CKINT)i;
-        }
-        else if( flags & Chuck_IO::INT16 || flags & Chuck_IO::UINT16 ) // 1.5.0.1 (ge) added UINT
-        {
-            // unsigned 16-bit
-            uint16_t i; // 1.5.0.1 (ge) changed to uint16_t from t_CKINT
-            m_io.read( (char *)&i, 2 );
-            if( m_io.gcount() != 2 )
-                EM_error3( "[chuck](via FileIO): cannot readInt: not enough bytes left" );
-            else if (m_io.fail())
-                EM_error3( "[chuck](via FileIO): cannot readInt: I/O stream failed" );
-            return (t_CKINT)i;
-        }
-        else if( flags & Chuck_IO::INT32 || flags & Chuck_IO::UINT32 ) // 1.5.0.1 (ge) added UINT
-        {
-            // unsigned 32-bit
-            uint32_t i; // 1.5.0.1 (ge) changed to uint32_t from t_CKINT
-            m_io.read( (char *)&i, 4 );
-            if (m_io.gcount() != 4)
-                EM_error3( "[chuck](via FileIO): cannot readInt: not enough bytes left" );
-            else if (m_io.fail())
-                EM_error3( "[chuck](via FileIO): cannot readInt: I/O stream failed" );
-            return (t_CKINT)i;
-        }
-        else if( flags & Chuck_IO::INT64 || flags & Chuck_IO::UINT64 ) // 1.5.0.1 (ge) added 64-bit
-        {
-            // unsigned 64-bit
-            uint64_t i; // 1.5.0.1 (ge) added
-            m_io.read( (char *)&i, 8 );
-            if (m_io.gcount() != 8)
-                EM_error3( "[chuck](via FileIO): cannot readInt: not enough bytes left" );
-            else if (m_io.fail())
-                EM_error3( "[chuck](via FileIO): cannot readInt: I/O stream failed" );
-            return (t_CKINT)i; // TODO: handle signed vs unsigned
-        }
-        else if( flags & Chuck_IO::SINT8 ) // 1.5.0.1 (ge) added SINT
-        {
-            // signed 8-bit
-            int8_t i;
-            m_io.read( (char *)&i, 1 );
-            if( m_io.gcount() != 1 )
-                EM_error3( "[chuck](via FileIO): cannot readInt: not enough bytes left" );
-            else if( m_io.fail() )
-                EM_error3( "[chuck](via FileIO): cannot readInt: I/O stream failed" );
-            return (t_CKINT)i;
-        }
-        else if( flags & Chuck_IO::SINT16 ) // 1.5.0.1 (ge) added SINT
-        {
-            // signed 16-bit
-            int16_t i;
-            m_io.read( (char *)&i, 2 );
-            if( m_io.gcount() != 2 )
-                EM_error3( "[chuck](via FileIO): cannot readInt: not enough bytes left" );
-            else if (m_io.fail())
-                EM_error3( "[chuck](via FileIO): cannot readInt: I/O stream failed" );
-            return (t_CKINT)i;
-        }
-        else if( flags & Chuck_IO::SINT32 ) // 1.5.0.1 (ge) added SINT
-        {
-            // signed 32-bit
-            int32_t i;
-            m_io.read( (char *)&i, 4 );
-            if (m_io.gcount() != 4)
-                EM_error3( "[chuck](via FileIO): cannot readInt: not enough bytes left" );
-            else if (m_io.fail())
-                EM_error3( "[chuck](via FileIO): cannot readInt: I/O stream failed" );
-            return (t_CKINT)i;
-        }
-        else if( flags & Chuck_IO::SINT64 ) // 1.5.0.1 (ge) added SINT
-        {
-            // signed 64-bit
-            int64_t i;
-            m_io.read( (char *)&i, 4 );
-            if (m_io.gcount() != 8)
-                EM_error3( "[chuck](via FileIO): cannot readInt: not enough bytes left" );
-            else if (m_io.fail())
-                EM_error3( "[chuck](via FileIO): cannot readInt: I/O stream failed" );
-            return (t_CKINT)i;
-        }
-        else
-        {
-            EM_error3( "[chuck](via FileIO): readInt error: invalid/unsupported int size flag" );
-            return 0;
-        }
-    }
-    else
-    {
-        EM_error3( "[chuck](via FileIO): readInt error: invalid ASCII/binary flag" );
-        return 0;
-    }
-}
-
-
-
-
-//-----------------------------------------------------------------------------
-// name: readFloat()
-// desc: read next as floating point value; could be ASCII or BINARY
-//-----------------------------------------------------------------------------
-t_CKFLOAT Chuck_IO_File::readFloat()
-{
-    return this->readFloat( Chuck_IO::FLOAT32 );
-}
-
-
-
-
-//-----------------------------------------------------------------------------
-// name: readFloat()
-// desc: read next as floating point value; could be ASCII or BINARY
-//-----------------------------------------------------------------------------
-t_CKFLOAT Chuck_IO_File::readFloat( t_CKINT flags )
-{
-    // sanity
-    if( !(m_io.is_open()) ) {
-        EM_error3( "[chuck](via FileIO): cannot readFloat: no file open" );
-        return 0;
-    }
-    if( m_io.eof() ) {
-        EM_error3( "[chuck](via FileIO): cannot readFloat: EOF reached" );
-        return 0;
-    }
-    if( m_io.fail() ) {
-        EM_error3( "[chuck](via FileIO): cannot readFloat: I/O stream failed" );
-        return 0;
-    }
-    if( m_dir ) {
-        EM_error3( "[chuck](via FileIO): cannot read a directory" );
-        return 0;
-    }
-
-    // check mode
-    if( m_flags & TYPE_ASCII )
-    {
-        // ASCII
-        t_CKFLOAT val = 0;
-        m_io >> val;
-        // if (m_io.fail())
-        //     EM_error3( "[chuck](via FileIO): cannot readFloat: I/O stream failed" );
-        return val;
-
-    }
-    else if( m_flags & TYPE_BINARY )
-    {
-        // 1.5.0.1
-        if( flags & Chuck_IO::FLOAT32 )
-        {
-            float i;
-            m_io.read( (char *)&i, sizeof(float) );
-            if (m_io.gcount() != sizeof(float) )
-                EM_error3( "[chuck](via FileIO): cannot readFloat: not enough bytes left" );
-            else if (m_io.fail())
-                EM_error3( "[chuck](via FileIO): cannot readFloat: I/O stream failed" );
-            return (t_CKFLOAT)i;
-        }
-        else if( flags & Chuck_IO::FLOAT64 )
-        {
-            double i;
-            m_io.read( (char *)&i, sizeof(double) );
-            if (m_io.gcount() != sizeof(double) )
-                EM_error3( "[chuck](via FileIO): cannot readFloat: not enough bytes left" );
-            else if (m_io.fail())
-                EM_error3( "[chuck](via FileIO): cannot readFloat: I/O stream failed" );
-            return (t_CKFLOAT)i;
-        }
-        else
-        {
-            EM_error3( "[chuck](via FileIO): readFloat error: invalid/unsupported datatype size flag" );
-            return 0;
-        }
-    }
-    else
-    {
-        EM_error3( "[chuck](via FileIO): readFloat error: invalid ASCII/binary flag" );
-        return 0;
-    }
-}
-
-
-
-
-//-----------------------------------------------------------------------------
-// name: readString()
-// desc: ...
-//-----------------------------------------------------------------------------
-t_CKBOOL Chuck_IO_File::readString( std::string & str )
-{
-    // set
-    str = "";
-
-    // sanity
-    if (!(m_io.is_open())) {
-        EM_error3( "[chuck](via FileIO): cannot readString: no file open" );
-        return FALSE;
-    }
-
-    if (m_io.eof()) {
-        EM_error3( "[chuck](via FileIO): cannot readString: EOF reached" );
-        return FALSE;
-    }
-
-    if ( m_dir )
-    {
-        EM_error3( "[chuck](via FileIO): cannot read on directory" );
-        return FALSE;
-    }
-
-    if (m_io.fail()) {
-        EM_error3( "[chuck](via FileIO): cannot readString: I/O stream failed" );
-        return FALSE;
-    }
-
-    if (m_flags & TYPE_ASCII) {
-        // ASCII
-        m_io >> str;
-        return TRUE;
-    } else if (m_flags & TYPE_BINARY) {
-        EM_error3( "[chuck](via FileIO): readString not supported for binary mode" );
-        return FALSE;
-    } else {
-        EM_error3( "[chuck](via FileIO): readInt error: invalid ASCII/binary flag" );
-        return FALSE;
-    }
-}
-
-
-
-
-/* (ATODO: doesn't look like asynchronous reads will work)
-
- THREAD_RETURN ( THREAD_TYPE Chuck_IO_File::read_thread ) ( void *data )
- {
- // (ATODO: test this)
- cerr << "In thread" << endl;
- async_args *args = (async_args *)data;
- Chuck_String *ret = (Chuck_String *)(args->RETURN);
- ret = args->fileio_obj->read( args->intArg );
- cerr << "Called FileIO.read(int)" << endl;
- args->fileio_obj->m_asyncEvent->broadcast(); // wake up
- cerr << "Broadcasted on event" << endl;
- cerr << "Sleeping" << endl;
- sleep(5);
- cerr << "Woke up" << endl;
- delete args;
- cerr << "Deleted args" << endl;
-
- return (THREAD_RETURN)0;
- }
-
- THREAD_RETURN ( THREAD_TYPE Chuck_IO_File::readLine_thread ) ( void *data )
- {
- // not yet implemented
- return NULL;
- }
-
- THREAD_RETURN ( THREAD_TYPE Chuck_IO_File::readInt_thread ) ( void *data )
- {
- // (ATODO: test this)
- async_args *args = (async_args *)data;
- Chuck_DL_Return *ret = (Chuck_DL_Return *)(args->RETURN);
- ret->v_int = args->fileio_obj->readInt( args->intArg );
- cerr << "Called readInt, set ret->v_int to " << ret->v_int << endl;
- args->fileio_obj->m_asyncEvent->broadcast(); // wake up
- cerr << "Called broadcast" << endl;
- delete args;
-
- return (THREAD_RETURN)0;
- }
-
- THREAD_RETURN ( THREAD_TYPE Chuck_IO_File::readFloat_thread ) ( void *data )
- {
- // not yet implemented
- return NULL;
- }
- */
-
-//-----------------------------------------------------------------------------
-// name: eof()
-// desc: end of file?
-//-----------------------------------------------------------------------------
-t_CKBOOL Chuck_IO_File::eof()
-{
-    if( !m_io.is_open() )
-    {
-        // EM_error3( "[chuck](via FileIO): cannot check eof: no file open" );
-        return TRUE;
-    }
-    if( m_dir )
-    {
-        EM_error3( "[chuck](via FileIO): cannot check eof on directory" );
-        return TRUE;
-    }
-
-    // return EOF conditions (1.5.0.0: peek() was added since eof() is set
-    // ONLY AFTER an effort is made to read and no more data is left)
-    // https://stackoverflow.com/questions/4533063/how-does-ifstreams-eof-work
-    return m_io.eof() || m_io.fail() || m_io.peek() == EOF;
-}
-
-
-
-
-//-----------------------------------------------------------------------------
-// name: write( const std::string & val )
-// desc: ...
-//-----------------------------------------------------------------------------
-void Chuck_IO_File::write( const std::string & val )
-{
-    // sanity
-    if (!(m_io.is_open())) {
-        EM_error3( "[chuck](via FileIO): cannot write: no file open" );
-        return;
-    }
-
-    if (m_io.fail()) {
-        EM_error3( "[chuck](via FileIO): cannot write: I/O stream failed" );
-        return;
-    }
-
-    if ( m_dir )
-    {
-        EM_error3( "[chuck](via FileIO): cannot write to a directory" );
-        return;
-    }
-
-    m_io.write( val.c_str(), val.size() );
-
-    if (m_io.fail()) { // check both before and after write if stream is ok
-        EM_error3( "[chuck](via FileIO): cannot write: I/O stream failed" );
-    }
-}
-
-
-
-
-//-----------------------------------------------------------------------------
-// name: write( t_CKINT val )
-// desc: ...
-//-----------------------------------------------------------------------------
-void Chuck_IO_File::write( t_CKINT val )
-{
-    // sanity
-    if (!(m_io.is_open())) {
-        EM_error3( "[chuck](via FileIO): cannot write: no file open" );
-        return;
-    }
-
-    if (m_io.fail()) {
-        EM_error3( "[chuck](via FileIO): cannot write: I/O stream failed" );
-        return;
-    }
-
-    if ( m_dir )
-    {
-        EM_error3( "[chuck](via FileIO): cannot write on directory" );
-        return;
-    }
-
-    if (m_flags & TYPE_ASCII) {
-        m_io << val;
-    } else if (m_flags & TYPE_BINARY) {
-        m_io.write( (char *)&val, sizeof(t_CKINT) );
-    } else {
-        EM_error3( "[chuck](via FileIO): write error: invalid ASCII/binary flag" );
-    }
-
-    if (m_io.fail()) { // check both before and after write if stream is ok
-        EM_error3( "[chuck](via FileIO): cannot write: I/O stream failed" );
-    }
-}
-
-
-
-
-//-----------------------------------------------------------------------------
-// name: write( t_CKINT val )
-// desc: ...
-//-----------------------------------------------------------------------------
-void Chuck_IO_File::write( t_CKINT val, t_CKINT flags )
-{
-    // sanity
-    if (!(m_io.is_open())) {
-        EM_error3( "[chuck](via FileIO): cannot write: no file open" );
-        return;
-    }
-
-    if (m_io.fail()) {
-        EM_error3( "[chuck](via FileIO): cannot write: I/O stream failed" );
-        return;
-    }
-
-    if ( m_dir )
-    {
-        EM_error3( "[chuck](via FileIO): cannot write on directory" );
-        return;
-    }
-
-    // check file mode
-    if( m_flags & TYPE_ASCII ) // ASCII mode
-    {
-        // insert into file stream
-        m_io << val;
-    }
-    else if( m_flags & TYPE_BINARY ) // BINARY mode
-    {
-        // check datatype size flags
-        // 1.5.0.1 (ge) modified to new signed/unsigned handling
-        if( flags & Chuck_IO::INT8 || flags & Chuck_IO::UINT8 )
-        {
-            // unsigned 8-bit
-            uint8_t v = val;
-            m_io.write( (char *)&v, 1 );
-        }
-        else if( flags & Chuck_IO::INT16 || flags & Chuck_IO::UINT16 )
-        {
-            // unsigned 16-bit
-            uint16_t v = val;
-            m_io.write( (char *)&v, 2 );
-        }
-        else if( flags & Chuck_IO::INT32 || flags & Chuck_IO::UINT32 )
-        {
-            // unsigned 32-bit
-            uint32_t v = (uint32_t)val;
-            m_io.write( (char *)&v, 4 );
-        }
-        else if( flags & Chuck_IO::INT64 || flags & Chuck_IO::UINT64 )
-        {
-            // unsigned 64-bit
-            uint64_t v = val;
-            m_io.write( (char *)&v, 8 );
-        }
-        else if( flags & Chuck_IO::SINT8 )
-        {
-            // signed 8-bit
-            int8_t v = val;
-            m_io.write( (char *)&v, 1 );
-        }
-        else if( flags & Chuck_IO::SINT16 )
-        {
-            // signed 16-bit
-            int16_t v = val;
-            m_io.write( (char *)&v, 2 );
-        }
-        else if( flags & Chuck_IO::SINT32 )
-        {
-            // signed 32-bit
-            int32_t v = (uint32_t)val;
-            m_io.write( (char *)&v, 4 );
-        }
-        else if( flags & Chuck_IO::SINT64 )
-        {
-            // signed 32-bit
-            int64_t v = val;
-            m_io.write( (char *)&v, 8 );
-        }
-    } else {
-        EM_error3( "[chuck](via FileIO): write error: invalid ASCII/binary flag" );
-    }
-
-    if (m_io.fail()) { // check both before and after write if stream is ok
-        EM_error3( "[chuck](via FileIO): cannot write: I/O stream failed" );
-    }
-}
-
-
-
-
-//-----------------------------------------------------------------------------
-// name: write( t_CKFLOAT val )
-// desc: write a floating point value
-//-----------------------------------------------------------------------------
-void Chuck_IO_File::write( t_CKFLOAT val )
-{
-    this->write( val, Chuck_IO::FLOAT32 );
-}
-
-
-
-
-//-----------------------------------------------------------------------------
-// name: write( t_CKFLOAT val, t_CKINT flags )
-// desc: write a floating point value; binary mode will heed flags for size
-//-----------------------------------------------------------------------------
-void Chuck_IO_File::write( t_CKFLOAT val, t_CKINT flags )
-{
-    // sanity
-    if( !(m_io.is_open()) ) {
-        EM_error3( "[chuck](via FileIO): cannot write: no file open" );
-        return;
-    }
-    if( m_io.fail() ) {
-        EM_error3( "[chuck](via FileIO): cannot write: I/O stream failed" );
-        return;
-    }
-    if( m_dir ) {
-        EM_error3( "[chuck](via FileIO): cannot write to a directory" );
-        return;
-    }
-
-    // check ASCII or BINARY
-    if( m_flags & TYPE_ASCII )
-    {
-        // insert into stream
-        m_io << val;
-    }
-    else if( m_flags & TYPE_BINARY )
-    {
-        // 1.5.0.1 (ge) add distinction between different float sizes
-        if( flags & Chuck_IO::FLOAT32 )
-        {
-            // 32-bit
-            float v = (float)val;
-            m_io.write( (char *)&v, 4 );
-        }
-        else if( flags & Chuck_IO::FLOAT64 )
-        {
-            // 64-bit
-            double v = (double)val;
-            m_io.write( (char *)&v, 8 );
-        }
-        else
-        {
-            EM_error3( "[chuck](via FileIO): writeFloat error: invalid/unsupport datatype size flag" );
-        }
-    }
-    else
-    {
-        EM_error3( "[chuck](via FileIO): write error: invalid ASCII/binary flag" );
-    }
-
-    if (m_io.fail()) { // check both before and after write if stream is ok
-        EM_error3( "[chuck](via FileIO): cannot write: I/O stream failed" );
-    }
-}
-
-
-
-
-#ifndef __DISABLE_THREADS__
-// static helper functions for writing asynchronously
-THREAD_RETURN ( THREAD_TYPE Chuck_IO_File::writeStr_thread ) ( void *data )
-{
-    async_args *args = (async_args *)data;
-    args->fileio_obj->write ( args->stringArg );
-    Chuck_Event *e = args->fileio_obj->m_asyncEvent;
-    delete args;
-    e->broadcast_local(); // wake up
-    e->broadcast_global();
-
-    return (THREAD_RETURN)0;
-}
-
-THREAD_RETURN ( THREAD_TYPE Chuck_IO_File::writeInt_thread ) ( void *data )
-{
-    async_args *args = (async_args *)data;
-    args->fileio_obj->write ( args->intArg );
-    args->fileio_obj->m_asyncEvent->broadcast_local(); // wake up
-    args->fileio_obj->m_asyncEvent->broadcast_global();
-    delete args;
-
-    return (THREAD_RETURN)0;
-}
-
-THREAD_RETURN ( THREAD_TYPE Chuck_IO_File::writeFloat_thread ) ( void *data )
-{
-    async_args *args = (async_args *)data;
-    args->fileio_obj->write( args->floatArg, args->intArg );
-    args->fileio_obj->m_asyncEvent->broadcast_local(); // wake up
-    args->fileio_obj->m_asyncEvent->broadcast_global();
-    delete args;
-
-    return (THREAD_RETURN)0;
-}
-#endif // __DISABLE_THREADS__
-// #endif // __DISABLE_FILEIO__
-
-
-
-
-//-----------------------------------------------------------------------------
-// name: Chuck_IO_Chout()
-// desc: constructor
-//-----------------------------------------------------------------------------
-Chuck_IO_Chout::Chuck_IO_Chout( Chuck_Carrier * carrier )
-{
-    // store
-    carrier->chout = this;
-    // add ref
-    this->add_ref();
-    // zero out
-    m_callback = NULL;
-    // initialize (added 1.3.0.0)
-    initialize_object( this, carrier->env->t_chout );
-    // lock so can't be deleted conventionally
-    this->lock();
-}
-
-Chuck_IO_Chout::~Chuck_IO_Chout()
-{
-    m_callback = NULL;
-}
-
-void Chuck_IO_Chout::set_output_callback( void (*fp)(const char *) )
-{
-    m_callback = fp;
-}
-
-t_CKBOOL Chuck_IO_Chout::good()
-{
-    return m_callback != NULL || cout.good();
-}
-
-void Chuck_IO_Chout::close()
-{ /* uh can't do it */ }
-
-void Chuck_IO_Chout::flush()
-{
-    if( m_callback )
-    {
-        // send to callback
-        m_callback( m_buffer.str().c_str() );
-    }
-    else
-    {
-        // print to cout
-        cout << m_buffer.str().c_str();
-        cout.flush();
-    }
-    // clear buffer
-    m_buffer.str( std::string() );
-}
-
-t_CKINT Chuck_IO_Chout::mode()
-{ return 0; }
-
-void Chuck_IO_Chout::mode( t_CKINT flag )
-{ }
-
-Chuck_String * Chuck_IO_Chout::readLine()
-{ return NULL; }
-
-t_CKINT Chuck_IO_Chout::readInt( t_CKINT flags )
-{ return 0; }
-
-t_CKFLOAT Chuck_IO_Chout::readFloat()
-{ return 0; }
-
-t_CKFLOAT Chuck_IO_Chout::readFloat( t_CKINT flags )
-{ return 0; }
-
-t_CKBOOL Chuck_IO_Chout::readString( std::string & str )
-{
-    str = "";
-    return FALSE;
-}
-
-t_CKBOOL Chuck_IO_Chout::eof()
-{ return TRUE; }
-
-void Chuck_IO_Chout::write( const std::string & val )
-// added 1.3.0.0: the flush
-{
-    m_buffer << val;
-    if( val == "\n" ) flush();
-}
-
-void Chuck_IO_Chout::write( t_CKINT val )
-{
-    m_buffer << val;
-}
-
-void Chuck_IO_Chout::write( t_CKINT val, t_CKINT flags )
-{
-    m_buffer << val;
-}
-
-void Chuck_IO_Chout::write( t_CKFLOAT val )
-{
-    m_buffer << val;
-}
-
-void Chuck_IO_Chout::write( t_CKFLOAT val, t_CKINT flags )
-{
-    // ignore flags for chout
-    m_buffer << val;
-}
-
-
-
-//-----------------------------------------------------------------------------
-// name: Chuck_IO_Cherr()
-// desc: constructor
-//-----------------------------------------------------------------------------
-Chuck_IO_Cherr::Chuck_IO_Cherr( Chuck_Carrier * carrier )
-{
-    // store
-    carrier->cherr = this;
-    // add ref
-    this->add_ref();
-    // zero out
-    m_callback = NULL;
-    // initialize (added 1.3.0.0)
-    initialize_object( this, carrier->env->t_cherr );
-    // lock so can't be deleted conventionally
-    this->lock();
-}
-
-Chuck_IO_Cherr::~Chuck_IO_Cherr()
-{
-    m_callback = NULL;
-}
-
-void Chuck_IO_Cherr::set_output_callback( void (*fp)(const char *) )
-{
-    m_callback = fp;
-}
-
-t_CKBOOL Chuck_IO_Cherr::good()
-{
-    return m_callback != NULL || cerr.good();
-}
-
-void Chuck_IO_Cherr::close()
-{ /* uh can't do it */ }
-
-void Chuck_IO_Cherr::flush()
-{
-    if( m_callback )
-    {
-        // send to callback
-        m_callback( m_buffer.str().c_str() );
-    }
-    else
-    {
-        // send to cerr
-        cerr << m_buffer.str().c_str();
-        cerr.flush();
-    }
-    // clear buffer
-    m_buffer.str( std::string() );
-}
-
-t_CKINT Chuck_IO_Cherr::mode()
-{ return 0; }
-
-void Chuck_IO_Cherr::mode( t_CKINT flag )
-{ }
-
-Chuck_String * Chuck_IO_Cherr::readLine()
-{ return NULL; }
-
-t_CKINT Chuck_IO_Cherr::readInt( t_CKINT flags )
-{ return 0; }
-
-t_CKFLOAT Chuck_IO_Cherr::readFloat()
-{ return 0; }
-
-t_CKFLOAT Chuck_IO_Cherr::readFloat( t_CKINT flags )
-{ return 0; }
-
-t_CKBOOL Chuck_IO_Cherr::readString( std::string & str )
-{
-    str = "";
-    return FALSE;
-}
-
-t_CKBOOL Chuck_IO_Cherr::eof()
-{ return TRUE; }
-
-void Chuck_IO_Cherr::write( const std::string & val )
-{
-    m_buffer << val;
-    flush(); // always flush for cerr | 1.5.0.0 (ge) added
-    // if( val == "\n" ) flush();
-}
-
-void Chuck_IO_Cherr::write( t_CKINT val )
-{
-    m_buffer << val;
-    flush(); // always flush for cerr | 1.5.0.0 (ge) added
-}
-
-void Chuck_IO_Cherr::write( t_CKINT val, t_CKINT flags )
-{
-    m_buffer << val;
-    flush(); // always flush for cerr | 1.5.0.0 (ge) added
-}
-
-void Chuck_IO_Cherr::write( t_CKFLOAT val )
-{
-    m_buffer << val;
-    flush(); // always flush for cerr | 1.5.0.0 (ge) added
-}
-
-void Chuck_IO_Cherr::write( t_CKFLOAT val, t_CKINT flags )
-{
-    // ignore flags for cherr
-    m_buffer << val;
-    flush(); // always flush for cerr | 1.5.0.0 (ge) added
 }
