+++ src/core/lo/lo_types_internal.h
@@ -13,7 +13,7 @@
 #include <poll.h>
 #endif
 
-#ifdef WIN32
+#ifdef _WIN32
 #include <winsock2.h>
 #include <ws2tcpip.h>
 #else
