+++ src/core/lo/message.c
@@ -20,7 +20,7 @@
 #include <string.h>
 #include <math.h>
 
-#ifdef WIN32
+#ifdef _WIN32
 #include <winsock2.h>
 #include <ws2tcpip.h>
 #else
