+++ src/core/lo/server.c
@@ -34,7 +34,7 @@
 #include <unistd.h>
 #endif
 
-#ifdef WIN32
+#ifdef _WIN32
 #include <winsock2.h>
 #include <ws2tcpip.h>
 #define EADDRINUSE WSAEADDRINUSE
