+++ src/core/ulib_opsc.cpp
@@ -35,6 +35,7 @@
 // lo.h ----> lo_osc_types.h needs to import <cstdint> before running itself,
 // but ONLY when compiling this file, and NOT when compiling liblo .c files
 // so, set a flag | REFACTOR-2017
+#include "chuck_def.h" // to pick up platform macros
 #ifdef __PLATFORM_WINDOWS__
 #define ULIB_OPSC_CPP
 #endif
@@ -45,16 +46,16 @@
 #include "lo/lo.h"
 #endif
 
+// on windows, the above must remain above these includes, else winsock errors
 #include "ulib_opsc.h"
+#include "chuck_compile.h"
+#include "chuck_instr.h"
 #include "chuck_type.h"
 #include "chuck_vm.h"
-#include "chuck_compile.h"
-#include "chuck_dl.h"
-#include "util_opsc.h"
-#include "chuck_instr.h"
-#include "util_thread.h"
 #include "util_buffers.h"
+#include "util_opsc.h"
 #include "util_string.h"
+#include "util_thread.h"
 
 
 #if _MSC_VER
